Development

+++ react-app/src/components/ProductDetailsPage/index.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from "react-redux";
-import { useParams, useHistory} from "react-router-dom";
+import { useParams} from "react-router-dom";
 
 
 import { productDetailsThunk } from '../../store/product';
@@ -23,6 +23,8 @@ function ProductDetailsPage() {
 
     const [ isLoaded, setIsLoaded ] = useState(false);
     const [ quantity, setQuantity ] = useState(1);
+    const [currentImageIndex, setCurrentImageIndex] = useState(0);
+
 
     const { openModal } = useModal()
     const [error, setError] = useState(null);
@@ -97,15 +99,21 @@ function ProductDetailsPage() {
 
 
     const handleDeleteReview = (reviewId) => {
-        dispatch(deleteReviewThunk(reviewId)).then(() => {
-            dispatch(getReviewThunk(productId));
-        });
+        const deteleIt = window.confirm("Are you sure you want to delete this review?");
+        if (deteleIt) {
+            dispatch(deleteReviewThunk(reviewId)).then(() => {
+                dispatch(getReviewThunk(productId));
+            });
+        }
     };
 
     const handleUpdateReview = (review) => {
         openModal(<UpdateReviewModal productId={productId} reviewId={review.id} />);
     };
 
+    const handleThumbnailClick = (index) => {
+        setCurrentImageIndex(index);
+    };
 
 
 
@@ -115,32 +123,28 @@ function ProductDetailsPage() {
 
             <div className="product-images-container">
                 <div className="thumbnail-images">
-                    {Array(6).fill(null).map((_, idx) => (
-                        <div className="thumbnail" key={idx}>
-                            {product.imageUrls && product.imageUrls.length > idx ? (
-                                <img src={product.imageUrls[idx]} alt={`${product.productName} Thumbnail ${idx + 1}`} />
-                            ) : (
-                                <div className="thumbnail-placeholder"></div>
-                            )}
+                    {product.imageUrls.map((imageUrl, idx) => (
+                        <div className="thumbnail" key={idx} onClick={() => handleThumbnailClick(idx)}>
+                            <img src={imageUrl} alt={`${product.productName} Thumbnail ${idx + 1}`} />
                         </div>
                     ))}
                 </div>
-                <div className="main-image">
-                    {product.imageUrls && product.imageUrls.length > 0 ?
-                        <img className="main-image-url" src={product.imageUrls[0]} alt={product.productName} />
-                        :
-                        // Placeholder for the main image when there are no images
-                        <div className="main-image-placeholder">No Image Available</div>
-                    }
-                </div>
-            </div>
 
+                <div className="main-image">
+                    {product.imageUrls && product.imageUrls[currentImageIndex] ? (
+                        <img className="main-image-url" src={product.imageUrls[currentImageIndex]} alt={product.productName} />
+                    ) : (
+                        <div className="main-image-placeholder">No Images Available</div>
+                    )}
+                </div>
+
+            </div>
 
             <div className="product-info">
                     <p className="detail-product-price">${product.price}</p>
 
                     <label>Quantity</label>
-                <div className="dropdown-container">
+                <div className="dropdown-container-details">
                         <select
                             className="quantity-dropdown"
                             value={quantity}
@@ -193,6 +197,7 @@ function ProductDetailsPage() {
             {reviews.map((review) => (
                 <li className="review-list-itself" key={review.id}>
                 <p className="reviews-username">{review.username}</p>
+                <p className="reviews-date">{new Date(review.dateCreated).toLocaleDateString()}</p>
 
                 <div className="reviews-rating">
                     {Array(5).fill(null).map((_, idx) => (
